feat: add verify if food exists

+++ src/services/FoodServices/DeleteFoodService.js
@@ -4,6 +4,12 @@ class DeleteFoodService {
   }
 
   async execute(id) {
+    const { food: foodExists } = await this.foodsRepository.findById(id)
+
+    if (!foodExists) {
+      throw new Error('Impossível de excluir! Prato inexistente!', 400)
+    }
+
     await this.foodsRepository.delete(id)
   }
 }
